refactor socketHandler
